+++ apps/sim/app/api/knowledge/[id]/documents/[documentId]/chunks/route.ts
@@ -2,8 +2,8 @@ import crypto from 'crypto'
 import { type NextRequest, NextResponse } from 'next/server'
 import { z } from 'zod'
 import { getSession } from '@/lib/auth'
-import { batchChunkOperation, createChunk, queryChunks } from '@/lib/chunks/service'
-import type { ChunkFilters, CreateChunkData } from '@/lib/chunks/types'
+import { batchChunkOperation, createChunk, queryChunks } from '@/lib/knowledge/chunks/service'
+import type { ChunkFilters, CreateChunkData } from '@/lib/knowledge/chunks/types'
 import { createLogger } from '@/lib/logs/console/logger'
 import { getUserId } from '@/app/api/auth/oauth/utils'
 import { checkDocumentAccess, checkDocumentWriteAccess } from '@/app/api/knowledge/utils'
