+++ apps/sim/app/api/knowledge/[id]/tag-definitions/route.ts
@@ -3,8 +3,8 @@ import { type NextRequest, NextResponse } from 'next/server'
 import { z } from 'zod'
 import { getSession } from '@/lib/auth'
 import { SUPPORTED_FIELD_TYPES } from '@/lib/constants/knowledge'
+import { createTagDefinition, getTagDefinitions } from '@/lib/knowledge/tags/service'
 import { createLogger } from '@/lib/logs/console/logger'
-import { createTagDefinition, getTagDefinitions } from '@/lib/tags/service'
 import { checkKnowledgeBaseAccess } from '@/app/api/knowledge/utils'
 
 export const dynamic = 'force-dynamic'
