+++ apps/sim/app/api/knowledge/[id]/tag-usage/route.ts
@@ -1,8 +1,8 @@
 import { randomUUID } from 'crypto'
 import { type NextRequest, NextResponse } from 'next/server'
 import { getSession } from '@/lib/auth'
+import { getTagUsage } from '@/lib/knowledge/tags/service'
 import { createLogger } from '@/lib/logs/console/logger'
-import { getTagUsage } from '@/lib/tags/service'
 import { checkKnowledgeBaseAccess } from '@/app/api/knowledge/utils'
 
 export const dynamic = 'force-dynamic'
