+++ apps/sim/lib/knowledge/documents/types.ts
