+++ apps/sim/lib/knowledge/documents/utils.ts
