+++ apps/sim/scripts/process-docs-embeddings.ts
@@ -2,8 +2,8 @@
 
 import path from 'path'
 import { sql } from 'drizzle-orm'
-import { DocsChunker } from '@/lib/documents/docs-chunker'
 import { isDev } from '@/lib/environment'
+import { DocsChunker } from '@/lib/knowledge/documents/docs-chunker'
 import { createLogger } from '@/lib/logs/console/logger'
 import { db } from '@/db'
 import { docsEmbeddings } from '@/db/schema'
